Removing the repositories and distributionManagement sections from the azure POM
During builds, the file-azure JAR was unable to deploy to the package registry, getting a 401 error. For example:

https://community.opengroup.org/osdu/platform/system/file/-/jobs/49231

Removing these lines should cause maven to use the parent pom's configuration, which is exactly the same. For my local build, that
caused it to work again. This commit will serve as a test case for the pipeline version.

I do not know why this caused the problem.